Rename pytorch_hooks.py to nvtx_pytorch_hooks.py
